# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Humain Numérique - Digital Human AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memory:
    get:
      operationId: listMemory
      tags: [memory]
      summary: List all memory entries
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Memory entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryEntry"
    delete:
      operationId: clearMemory
      tags: [memory]
      summary: Clear all memory
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /memory/search:
    get:
      operationId: searchMemory
      tags: [memory]
      summary: Search memory entries
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryEntry"
...
  /files/pdf:
    post:
      operationId: generatePdf
      tags: [files]
      summary: Generate a PDF document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PdfInput"
      responses:
        "200":
          description: PDF file info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileResult"
  /files/zip:
    post:
      operationId: generateZip
      tags: [files]
      summary: Generate a ZIP archive
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZipInput"
      responses:
        "200":
          description: ZIP file info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileResult"
...
  /vercel/deploy:
    post:
      operationId: deployToVercel
      tags: [vercel]
      summary: Deploy code to Vercel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VercelDeployInput"
      responses:
        "200":
          description: Deployment info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VercelDeployment"
  /vercel/deployments/{id}:
    get:
      operationId: getVercelDeployment
      tags: [vercel]
      summary: Get Vercel deployment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deployment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VercelDeployment"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    MemoryEntry:
      type: object
      required: [id, prompt, response, createdAt]
      properties:
        id:
          type: string
        prompt:
          type: string
        response:
          type: string
        model:
          type: string
        accessCount:
          type: integer
        createdAt:
          type: string
...
    PdfInput:
      type: object
      required: [content, title]
      properties:
        title:
          type: string
        content:
          type: string
        format:
          type: string
          enum: [A4, letter]
          default: A4
    ZipInput:
      type: object
      required: [files, name]
      properties:
        name:
          type: string
        files:
          type: array
          items:
            type: object
            required: [filename, content]
            properties:
              filename:
                type: string
              content:
                type: string
    FileResult:
      type: object
      required: [id, name, type, downloadUrl, size, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [pdf, zip]
        downloadUrl:
          type: string
        size:
          type: integer
        createdAt:
          type: string
...
    VercelDeployInput:
      type: object
      required: [name, files]
      properties:
        name:
          type: string
        files:
          type: array
          items:
            type: object
            required: [file, data]
            properties:
              file:
                type: string
              data:
                type: string
    VercelDeployment:
      type: object
      required: [id, url, state, createdAt]
      properties:
        id:
          type: string
        url:
          type: string
        state:
          type: string
          enum: [BUILDING, READY, ERROR, CANCELED]
        name:
          type: string
        createdAt:
          type: string
        inspectorUrl:
          type: ["string", "null"]
...